perf(doctor): avoid loading unused plugin runtimes
